+++ assets/js/blocks/breeze-blocks.asset.php
@@ -6,5 +6,5 @@ return array(
         'wp-element',
         'wp-html-entities',
     ),
-    'version' => '2.0.0',
+    'version' => '2.0.1',
 );